requirements.txt
python-dotenv
psycopg2-binary
playwright
pinecone[asyncio]
apache-airflow
apache-airflow-providers-postgres